+++ pkg/tunnel/pool.go
@@ -5,6 +5,9 @@ import (
 	"errors"
 	"sync"
 
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+
 	"github.com/datawire/dlib/dlog"
 )
 
@@ -56,8 +59,12 @@ func (p *Pool) GetOrCreate(ctx context.Context, id ConnID, createHandler Handler
 	}
 
 	handlerCtx, cancel := context.WithCancel(ctx)
+	handlerCtx, span := otel.Tracer("").Start(handlerCtx, "handler")
+	span.SetAttributes(attribute.String("conn-id", id.String()))
+
 	release := func() {
 		p.release(ctx, id)
+		span.End()
 		cancel()
 	}